// src/app/visitantes/models/interfaces/visitante.interfaces.ts
// Generated by https://quicktype.io

export interface IVisitante {
    apellidos: string;
    email: string;
    id: string;
    nombre: string;
    numeroAcompanantes: number;
}

export interface IVisitanteRequest {
    apellidos: string;
    email: string;
    nombre: string;
    numeroAcompanantes: number;
}